Move zipdatadir to common-ts, move validation fns to common-ts
