+++ common-ts/lib/index.ts
@@ -4,10 +4,11 @@ import { ServerMessage } from './type/server-message'
 import { MessageType, ServerResponse, VersionResponse } from './type/server-types'
 
 import { getMessageBody } from './util/message-util'
+import { validateString } from './util/validation-util'
 
 
-let SERVER_URL = 'http://localhost:9999/'
-let AUTHORIZATION_HEADER = ''
+let SERVER_URL = 'http://localhost:9999/' // default to bitcoin-s server
+let AUTHORIZATION_HEADER = '' // default to no auth
 
 /** Set Wallet Server endpoint */
 export function ConfigureServerURL(url: string) {
@@ -18,6 +19,11 @@ export function ConfigureServerURL(url: string) {
 export function ConfigureAuthorizationHeader(header: string) {
   console.debug('ConfigureAuthorizationHeader()', header)
   AUTHORIZATION_HEADER = header
+}
+// Convenience function
+export function ConfigureAuthorizationHeaderFromUserPassword(user: string, password: string) {
+  console.debug('ConfigureAuthorizationHeader()', user)
+  AUTHORIZATION_HEADER = 'Basic ' + Buffer.from(`${user}:${password}`).toString('base64')
 }
 
 /** Send any ServerMessage */
@@ -39,7 +45,7 @@ export function SendServerMessage(message: ServerMessage) {
   }
 }
 
-/** Specific bitcoin-s Server message functions */ 
+/** Common bitcoin-s message functions */ 
 
 export function GetVersion() {
   console.debug('GetVersion()')
@@ -47,6 +53,17 @@ export function GetVersion() {
   const m = getMessageBody(MessageType.getversion)
   return SendServerMessage(m).then(response => {
     return <ServerResponse<VersionResponse>>response
+  })
+}
+
+export function ZipDataDir(path: string) {
+  console.debug('ZipDataDir()')
+  validateString(path, 'ZipDataDir()', 'path')
+
+  const m = getMessageBody(MessageType.zipdatadir, [path])
+  return SendServerMessage(m).then(response => {
+    // result: 'failure' / null
+    return <ServerResponse<string|null>>response
   })
 }
