--- oracle-server-ts/lib/type/oracle-server-message.ts
@@ -1,19 +1,0 @@
-import { MessageType } from './oracle-server-types'
-
-
-export class OracleServerMessage {
-  method: MessageType
-
-  constructor(type: MessageType) {
-    this.method = type
-  }
-}
-
-export class OracleServerMessageWithParameters extends OracleServerMessage {
-  params: any[] // usually string[]
-
-  constructor(type: MessageType, params: any[]) {
-    super(type)
-    this.params = params
-  }
-}
