+++ oracle-server-ts/lib/util/validation-util.ts
@@ -1,20 +1,4 @@
 
-export function validateString(s: string, fromFn: string, paramName: string) {
-  if (typeof s !== 'string') {
-    throw(Error(fromFn + ' non-string ' + paramName))
-  }
-}
-
-const ISO_DATE_REGEX = /[+-]?\d{4}(-[01]\d(-[0-3]\d(T[0-2]\d:[0-5]\d:?([0-5]\d(\.\d+)?)?[+-][0-2]\d:[0-5]\dZ?)?)?)?/
-
-export function validateISODateString(s: string, fromFn: string, paramName: string) {
-  validateString(s, fromFn, paramName)
-  // TODO : Validate ISO Date String
-  if (!ISO_DATE_REGEX.test(s)) {
-    throw(Error(fromFn + 'ISO Date invalid'))
-  }
-}
-
 export function validateEnumOutcomes(outcomes: string[], fromFn: string) {
   if (outcomes && outcomes.length > 1) {
     const hasEmpty = outcomes.filter(i => i !== '')
@@ -29,12 +13,6 @@ export function validateEnumOutcomes(outcomes: string[], fromFn: string) {
     throw(Error(fromFn + ' must have at least two outcomes'))
   } else {
     throw(Error(fromFn + ' outcomes invalid'))
-  }
-}
-
-export function validateNumber(n: number, fromFn: string, paramName: string) {
-  if (typeof n !== 'number') {
-    throw(Error(fromFn + ' non-number ' + paramName))
   }
 }
